server: move the grpc gateway init code out of server.go
Release note: None
