+++ pkg/server/BUILD.bazel
@@ -22,6 +22,7 @@ go_library(
         "drain.go",
         "env_sampler.go",
         "external_storage_builder.go",
+        "grpc_gateway.go",
         "grpc_server.go",
         "gzip_response_writer.go",
         "import_ts.go",
